+++ routes/index.js
@@ -31,7 +31,6 @@ if (db === "") {
 //Dashboard Page
 router.get("/dashboard", ensureAuthenticated, async (req, res) => {
   let printers = await Runner.returnFarmPrinters();
-  printers = printers.sort((a, b) => (a.sortIndex > b.sortIndex ? 1 : -1));
   const farmStatistics = require("../runners/statisticsCollection.js");
   const FarmStatistics = farmStatistics.StatisticsCollection;
   let statistics = await FarmStatistics.returnStats();
@@ -45,6 +44,7 @@ router.get("/dashboard", ensureAuthenticated, async (req, res) => {
   }
   res.render("dashboard", {
     name: user,
+    userGroup: req.user.group,
     version: pjson.version,
     printers: printers,
     farmInfo: statistics.farmInfo,
@@ -62,7 +62,6 @@ router.get("/dashboard", ensureAuthenticated, async (req, res) => {
 //File Manager Page
 router.get("/filemanager", ensureAuthenticated, async (req, res) => {
   let printers = await Runner.returnFarmPrinters();
-  printers = printers.sort((a, b) => (a.sortIndex > b.sortIndex ? 1 : -1));
   let filament = await Filament.find({});
   const farmStatistics = require("../runners/statisticsCollection.js");
   const FarmStatistics = farmStatistics.StatisticsCollection;
@@ -74,6 +73,7 @@ router.get("/filemanager", ensureAuthenticated, async (req, res) => {
   }
   res.render("filemanager", {
     name: user,
+    userGroup: req.user.group,
     version: pjson.version,
     printers: printers,
     printerCount: printers.length,
@@ -86,7 +86,6 @@ router.get("/filemanager", ensureAuthenticated, async (req, res) => {
 //History Page
 router.get("/history", ensureAuthenticated, async (req, res) => {
   let printers = Runner.returnFarmPrinters();
-  printers = printers.sort((a, b) => (a.sortIndex > b.sortIndex ? 1 : -1));
   const History = require("../models/History.js");
   let history = await History.find({});
   let filament = await Filament.find({});
@@ -98,6 +97,7 @@ router.get("/history", ensureAuthenticated, async (req, res) => {
   }
   res.render("history", {
     name: user,
+    userGroup: req.user.group,
     version: pjson.version,
     printers: printers,
     printerCount: printers.length,
@@ -111,7 +111,6 @@ router.get("/history", ensureAuthenticated, async (req, res) => {
 //Panel view  Page
 router.get("/mon/panel", ensureAuthenticated, async (req, res) => {
   let printers = Runner.returnFarmPrinters();
-  printers = printers.sort((a, b) => (a.sortIndex > b.sortIndex ? 1 : -1));
   const farmStatistics = require("../runners/statisticsCollection.js");
   const FarmStatistics = farmStatistics.StatisticsCollection;
   let statistics = await FarmStatistics.returnStats();
@@ -125,6 +124,7 @@ router.get("/mon/panel", ensureAuthenticated, async (req, res) => {
   }
   res.render("panelView", {
     name: user,
+    userGroup: req.user.group,
     version: pjson.version,
     printers: printers,
     currentOperations: statistics.currentOperations,
@@ -140,7 +140,6 @@ router.get("/mon/panel", ensureAuthenticated, async (req, res) => {
 //Camera view  Page
 router.get("/mon/camera", ensureAuthenticated, async (req, res) => {
   let printers = Runner.returnFarmPrinters();
-  printers = printers.sort((a, b) => (a.sortIndex > b.sortIndex ? 1 : -1));
   const farmStatistics = require("../runners/statisticsCollection.js");
   const FarmStatistics = farmStatistics.StatisticsCollection;
   let statistics = await FarmStatistics.returnStats();
@@ -154,6 +153,7 @@ router.get("/mon/camera", ensureAuthenticated, async (req, res) => {
   }
   res.render("cameraView", {
     name: user,
+    userGroup: req.user.group,
     version: pjson.version,
     printers: printers,
     currentOperations: statistics.currentOperations,
@@ -169,7 +169,6 @@ router.get("/mon/camera", ensureAuthenticated, async (req, res) => {
 //List view  Page
 router.get("/mon/list", ensureAuthenticated, async (req, res) => {
   let printers = Runner.returnFarmPrinters();
-  printers = printers.sort((a, b) => (a.sortIndex > b.sortIndex ? 1 : -1));
   const farmStatistics = require("../runners/statisticsCollection.js");
   const FarmStatistics = farmStatistics.StatisticsCollection;
   let statistics = await FarmStatistics.returnStats();
@@ -183,6 +182,7 @@ router.get("/mon/list", ensureAuthenticated, async (req, res) => {
   }
   res.render("listView", {
     name: user,
+    userGroup: req.user.group,
     version: pjson.version,
     printers: printers,
     currentOperations: statistics.currentOperations,